+++ .gitignore
@@ -29,4 +29,3 @@ yarn-error.log*
 
 # Optional eslint cache
 .eslintcache
-.env
