+++ package.json
@@ -5,8 +5,8 @@
   "type": "module",
   "homepage": "https://jemjem7.github.io/it35-lab",
   "scripts": {
-    "predeploy": "npm run build",
-    "deploy": "gh-pages -d dist",
+    "predeploy":"npm run build",
+    "deploy":"gh-pages -d dist",
     "dev": "vite",
     "build": "tsc && vite build",
     "preview": "vite preview",
@@ -22,11 +22,8 @@
     "@capacitor/status-bar": "7.0.0",
     "@ionic/react": "^8.0.0",
     "@ionic/react-router": "^8.0.0",
-    "@supabase/supabase-js": "^2.49.4",
     "@types/react-router": "^5.1.20",
     "@types/react-router-dom": "^5.3.3",
-    "animate.css": "^4.1.1",
-    "bcryptjs": "^3.0.2",
     "ionicons": "^7.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
