+++ src/pages/Menu.tsx
@@ -1,5 +1,4 @@
 import { 
-    IonAlert,
     IonButton,
     IonButtons,
       IonContent, 
@@ -13,43 +12,19 @@ import {
       IonRouterOutlet, 
       IonSplitPane, 
       IonTitle, 
-      IonToast, 
-      IonToolbar, 
-      useIonRouter
+      IonToolbar 
   } from '@ionic/react'
-  import {homeOutline, logOutOutline, rocketOutline, settingsOutline} from 'ionicons/icons';
+  import {homeOutline, logOutOutline, rocketOutline} from 'ionicons/icons';
 import { Redirect, Route } from 'react-router';
 import Home from './Home';
 import About from './About';
 import Details from './Details';
-import { supabase } from '../utils/supabaseClient';
-import { useState } from 'react';
-import EditProfilePage from './EditProfilePage';
-
 
   const Menu: React.FC = () => {
-    const navigation = useIonRouter();
-    const [showAlert, setShowAlert] = useState(false);
-    const [errorMessage, setErrorMessage] = useState('');
-    const [showToast, setShowToast] = useState(false);
-    
     const path = [
         {name:'Home', url: '/it35-lab/app/home', icon: homeOutline},
         {name:'About', url: '/it35-lab/app/about', icon: rocketOutline},
-        {name:'Profile', url: '/it35-lab/app/profile', icon: settingsOutline},
     ]
-    const handleLogout = async () => {
-        const { error } = await supabase.auth.signOut();
-        if (!error) {
-            setShowToast(true);
-            setTimeout(() => {
-                navigation.push('/it35-lab', 'back', 'replace'); 
-            }, 300); 
-        } else {
-            setErrorMessage(error.message);
-            setShowAlert(true);
-        }
-    };
 
     return (
         <IonPage>
@@ -72,10 +47,10 @@ import EditProfilePage from './EditProfilePage';
                             </IonMenuToggle>
                         ))}
 
-                       {/* Logout Button */}
-                       <IonButton expand="full" onClick={handleLogout}>
-                            <IonIcon icon={logOutOutline} slot="start"></IonIcon>
-                            Logout
+                        {/*Logout Button*/}
+                        <IonButton routerLink="/it35-lab" routerDirection="back" expand="full">
+                            <IonIcon icon={logOutOutline} slot="start"> </IonIcon>
+                        Logout
                         </IonButton>
                         
                     </IonContent>
@@ -85,31 +60,11 @@ import EditProfilePage from './EditProfilePage';
                     <Route exact path="/it35-lab/app/home" component={Home} />
                     <Route exact path="/it35-lab/app/home/details" component={Details} />
                     <Route exact path="/it35-lab/app/about" component={About} />
-                    <Route exact path="/it35-lab/app/profile" component={EditProfilePage} />
 
                     <Route exact path="/it35-lab/app">
                         <Redirect to="/it35-lab/app/home"/>
                     </Route>
                 </IonRouterOutlet>
-
-                {/* IonAlert for displaying login errors */}
-                <IonAlert
-                    isOpen={showAlert}
-                    onDidDismiss={() => setShowAlert(false)}
-                    header="Logout Failed"
-                    message={errorMessage}
-                    buttons={['OK']}
-                />
-                
-                {/* IonToast for success message */}
-                <IonToast
-                    isOpen={showToast}
-                    onDidDismiss={() => setShowToast(false)}
-                    message="Logout Successful"
-                    duration={1500}
-                    position="top"
-                    color="primary"
-                />
 
             </IonSplitPane>
         </IonPage>
