+++ src/pages/home-tabs/Feed.tsx
@@ -1,37 +1,38 @@
 import { 
-  IonButtons,
-    IonCard,
-    IonCardContent,
-    IonCardHeader,
-    IonCardSubtitle,
-    IonCardTitle,
-    IonContent, 
-    IonHeader, 
-    IonMenuButton, 
-    IonPage, 
-    IonSearchbar, 
-    IonTitle, 
-    IonToolbar 
-} from '@ionic/react';
-import FeedContainer from '../../components/FeedContainer';
-
-const Feed: React.FC = () => {
-  return (
-    <IonPage>
-      <IonHeader>
-        <IonToolbar>
-          <IonButtons slot='start'>
-            <IonMenuButton></IonMenuButton>
-          </IonButtons>
-          <IonTitle>Feed</IonTitle>
-        </IonToolbar>
-      </IonHeader>
-      <IonContent fullscreen>
-      
-    <FeedContainer />
-      </IonContent>
-    </IonPage>
-  );
-};
-
-export default Feed;
+    IonButtons,
+      IonContent, 
+      IonHeader, 
+      IonMenuButton, 
+      IonPage, 
+      IonTitle, 
+      IonToolbar 
+  } from '@ionic/react';
+  
+  const Feed: React.FC = () => {
+    return (
+      <IonPage>
+        <IonHeader>
+          <IonToolbar>
+            <IonButtons slot='start'>
+              <IonMenuButton></IonMenuButton>
+            </IonButtons>
+            <IonTitle>Feed</IonTitle>
+          </IonToolbar>
+        </IonHeader>
+        <IonContent fullscreen>
+          <div
+            style={{
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'center',
+              height: '100%',
+            }}
+          >
+            Feed
+          </div>
+        </IonContent>
+      </IonPage>
+    );
+  };
+  
+  export default Feed;
