+++ src/pages/home-tabs/Search.tsx
@@ -1,49 +1,39 @@
 import { 
-  IonButtons,
-    IonContent, 
-    IonHeader, 
-    IonMenuButton, 
-    IonPage, 
-    IonTitle, 
-    IonToolbar, 
-    IonSearchbar,
-    IonGrid,
-    IonRow,
-    IonCol,
-
-} from '@ionic/react';
-
-const Search: React.FC = () => {
-  return (
-    <IonPage>
-      <IonHeader>
-        <IonToolbar>
-          <IonButtons slot='start'>
-            <IonMenuButton></IonMenuButton>
-          </IonButtons>
-          <IonTitle>Search</IonTitle>
-        </IonToolbar>
-      </IonHeader>
-      <IonContent fullscreen>
-         
-      <IonGrid>
-      <IonRow className="ion-justify-content-center">
-
-            <IonCol size="12" sizeMd="8" sizeLg="6" className="ion-padding">
-            <IonSearchbar 
-              placeholder="Search"
-              debounce={0} // Optional: Optional debounce for input delay
-              showClearButton="focus"
-              
-       
-               />
-             </IonCol>
-          </IonRow>
-        </IonGrid>
-       
-      </IonContent>
-    </IonPage>
-  );
-};
-
-export default Search;
+    IonButtons,
+      IonContent, 
+      IonHeader, 
+      IonMenuButton, 
+      IonPage, 
+      IonTitle, 
+      IonToolbar 
+  } from '@ionic/react';
+  
+  const Search: React.FC = () => {
+    return (
+      <IonPage>
+        <IonHeader>
+          <IonToolbar>
+            <IonButtons slot='start'>
+              <IonMenuButton></IonMenuButton>
+            </IonButtons>
+            <IonTitle>Search</IonTitle>
+          </IonToolbar>
+        </IonHeader>
+        <IonContent fullscreen>
+          <div
+            style={{
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'center',
+              height: '100%',
+            }}
+          >
+            Search
+          </div>
+  
+        </IonContent>
+      </IonPage>
+    );
+  };
+  
+  export default Search;
